Check the validity of arguments passed to multi_gene.R plotting functions

Also subset spe to the selected sample and genes, dropping genes with constant expression across spots

.multi_gene_validity_check(spe, genes, sample_id, assayname, minCount, ...)

Arguments

spe

A SpatialExperiment with colData column exclude_overlapping, passed to spatialLIBD::vis_gene or spatialLIBD::vis_clus

genes

character() of length >= 2 of gene names to plot in combination, expected to be in rownames(spe)

sample_id

character(1) passed to sampleid in spatialLIBD::vis_gene or spatialLIBD::vis_clus. Assumed to be a donor, possibly consisting of several capture areas to plot at once

assayname

character(1) passed to spatialLIBD::vis_gene

minCount

numeric(1) passed to passed to spatialLIBD::vis_gene

...

Parameters accepted by spot_plot, excluding is_discrete or var_name, which are handled internally

Value

SpatialExperiment subsetted to the specified sample and to each of the non-constant-expression genes

Author

Nicholas J. Eagles